requirements.txt
playwright>=1.40.0
openai>=1.3.0
opencv-python>=4.8.1.78
pytest>=7.4.3
pytest-asyncio>=0.21.1
requests>=2.31.0
locust>=2.17.0
Pillow>=10.1.0
PyYAML>=6.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
selenium>=4.15.2
numpy>=1.24.3